Remove Date extension method with translatons needed and same for DeviceInfo and moved those to Localizations framework together with Strings.swift file
