+++ BrowserKit/Sources/Shared/TimeConstants.swift
@@ -85,61 +85,6 @@ extension Date {
         return Int((currentYear / 1000) * 1000)
     }
 
-    public func toRelativeTimeString(
-        dateStyle: DateFormatter.Style = .short,
-        timeStyle: DateFormatter.Style = .short
-    ) -> String {
-        let now = Date()
-
-        let units: Set<Calendar.Component> = [.second, .minute, .day, .weekOfYear, .month, .year, .hour]
-        let components = Calendar.current.dateComponents(units, from: self, to: now)
-
-        if components.year ?? 0 > 0 {
-            return String(
-                format: DateFormatter.localizedString(
-                    from: self,
-                    dateStyle: dateStyle,
-                    timeStyle: timeStyle
-                )
-            )
-        }
-
-        if components.month == 1 {
-            return String(format: .TimeConstantMoreThanAMonth)
-        }
-
-        if components.month ?? 0 > 1 {
-            return String(
-                format: DateFormatter.localizedString(
-                    from: self,
-                    dateStyle: dateStyle,
-                    timeStyle: timeStyle
-                )
-            )
-        }
-
-        if components.weekOfYear ?? 0 > 0 {
-            return String(format: .TimeConstantMoreThanAWeek)
-        }
-
-        if components.day == 1 {
-            return String(format: .TimeConstantYesterday)
-        }
-
-        if components.day ?? 0 > 1 {
-            return String(format: .TimeConstantThisWeek, String(describing: components.day))
-        }
-
-        if components.hour ?? 0 > 0 || components.minute ?? 0 > 0 {
-            // Can't have no time specified for this formatting case.
-            let timeStyle = timeStyle != .none ? timeStyle : .short
-            let absoluteTime = DateFormatter.localizedString(from: self, dateStyle: .none, timeStyle: timeStyle)
-            return String(format: .TimeConstantRelativeToday, absoluteTime)
-        }
-
-        return String(format: .TimeConstantJustNow)
-    }
-
     public func toRFC822String() -> String {
         return rfc822DateFormatter.string(from: self)
     }
