+++ firefox-ios/Shared/Strings.swift
