+++ firefox-ios/firefox-ios-tests/Tests/ClientTests/Nimbus/NimbusIntegrationTests.swift
@@ -3,7 +3,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 import Foundation
-@testable import Shared
+@testable import Localizations
 
 import XCTest
